requirements.txt
asgiref==3.3.0
astroid==2.4.2
backcall==0.2.0
decorator==4.4.2
Django==3.1.3
django-extensions==3.0.9
djangorestframework==3.12.2
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
pylint==2.6.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
wrapt==1.12.1
